Version 1.0.71
https://www.gurux.fi/node/15787

+++ Gurux.DLMS.python/gurux_dlms/GXUInt32.py
@@ -31,5 +31,5 @@
 #  This code is licensed under the GNU General Public License v2.
 #  Full text may be retrieved at http://www.gnu.org/licenses/gpl-2.0.txt
 # ---------------------------------------------------------------------------
-class GXUInt32(int):
+class GXUInt32(long):
     """UInt32 class."""
